pyproject.toml
[project]
name = "capital-allocation-strategy-agent"
version = "0.1.0"
description = "A multi-agent system that automates capital allocation strategy with human-in-the-loop approval and repair & resume support."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = "Apache-2.0"
authors = [
    { name = "Nora Filali", email = "nora.filali@google.com" },
]
maintainers = [
    { name = "Nora Filali", email = "nora.filali@google.com" },
]
dependencies = [
    "google-adk>=1.27.2",
    "google-cloud-firestore>=2.21.0",
    "reportlab>=4.4.10",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "google-adk[eval]>=1.27.2",
    "nest-asyncio>=1.6.0",
    "agent-starter-pack>=0.14.1",
    "ipykernel>=6.29.0",
]

[project.optional-dependencies]
lint = [
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""

[tool.agent-starter-pack]
example_question = "We are losing 4% MoM in the 'Artisan' segment to Square/Stripe. Analyze our Q1 risk data and external market comps. Identify the 'High-Growth' segment we can win back and propose a proactive capital structure that beats the competition while staying within our Tier 1 capital ratios."

[tool.agent-starter-pack.settings]
agent_directory = "capital_allocation_strategy_agent"
